Bar Chart page added
