--- app1.py
@@ -1,4 +1,0 @@
-import streamlit as st
-def app():
-    st.title('APP1')
-    st.write('Welcome to app1')
