+++ bar.py
@@ -1,0 +1,33 @@
+import streamlit as st
+import altair as alt
+import pandas as pd
+from add_data import db_execute_fetch
+
+def app():
+    st.title('Bar Chart')
+    st.write('You can see Ranking here')
+    stBarChart()
+
+def loadData():
+    query = "select * from TweetInformation"
+    df = db_execute_fetch(query, dbName="tweets", rdf=True)
+    return df
+
+def barChart(data, title, X, Y):
+    title = title.title()
+    st.title(f'{title} Chart')
+    msgChart = (alt.Chart(data).mark_bar().encode(alt.X(f"{X}:N", sort=alt.EncodingSortField(field=f"{Y}", op="values",
+                order='ascending')), y=f"{Y}:Q"))
+    st.altair_chart(msgChart, use_container_width=True)
+
+def stBarChart():
+    df = loadData()
+    dfCount = pd.DataFrame({'Tweet_count': df.groupby(['original_author'])['clean_text'].count()}).reset_index()
+    dfCount["original_author"] = dfCount["original_author"].astype(str)
+    dfCount = dfCount.sort_values("Tweet_count", ascending=False)
+
+    num = st.slider("Select number of Rankings", 0, 50, 5)
+    title = f"Top {num} Ranking By Number of tweets"
+    barChart(dfCount.head(num), title, "original_author", "Tweet_count")
+
+
