+++ streamlit.py
@@ -7,10 +7,10 @@ import plotly.express as px
 from add_data import db_execute_fetch
 st.set_page_config(page_title="Dashboard", layout="wide")
 
-import app1
+import bar
 import app2
 PAGES = {
-    "App1": app1,
+    "Bar Chart ": bar,
     "App2": app2
 }
 st.sidebar.title('Pages')
@@ -50,12 +50,7 @@ def selectLocAndAuth():
     else:
         st.write(df)
 
-def barChart(data, title, X, Y):
-    title = title.title()
-    st.title(f'{title} Chart')
-    msgChart = (alt.Chart(data).mark_bar().encode(alt.X(f"{X}:N", sort=alt.EncodingSortField(field=f"{Y}", op="values",
-                order='ascending')), y=f"{Y}:Q"))
-    st.altair_chart(msgChart, use_container_width=True)
+
 
 def wordCloud():
     df = loadData()
@@ -69,15 +64,6 @@ def wordCloud():
     st.title("Tweet Text Word Cloud")
     st.image(wc.to_array())
 
-def stBarChart():
-    df = loadData()
-    dfCount = pd.DataFrame({'Tweet_count': df.groupby(['original_author'])['clean_text'].count()}).reset_index()
-    dfCount["original_author"] = dfCount["original_author"].astype(str)
-    dfCount = dfCount.sort_values("Tweet_count", ascending=False)
-
-    num = st.slider("Select number of Rankings", 0, 50, 5)
-    title = f"Top {num} Ranking By Number of tweets"
-    barChart(dfCount.head(num), title, "original_author", "Tweet_count")
 
 
 def langPie():
@@ -111,5 +97,5 @@ selectLocAndAuth()
 st.title("Data Visualizations")
 wordCloud()
 with st.beta_expander("Show More Graphs"):
-    stBarChart()
+
     langPie()
